@@ -1,20 +1,23 @@
 {
   "name": "quasar-ui-linkage-cms-ui",
-  "version": "1.6.0",
+  "version": "2.0.0",
   "author": "Robby <ta7382@gmail.com>",
   "description": "Linkage CMS UI",
   "license": "MIT",
-  "module": "src/index.esm.js",
-  "main": "src/index.common.js",
+  "module": "dist/index.esm.js",
+  "main": "dist/index.common.js",
   "scripts": {
     "dev": "cd dev && yarn dev && cd ..",
     "dev:umd": "yarns build && node build/script.open-umd.js",
-    "dev:ssr": "cd dev && yarn 'dev:ssr' && cd ..",
-    "dev:ios": "cd dev && yarn 'dev:ios' && cd ..",
-    "dev:android": "cd dev && yarn 'dev:android' && cd ..",
-    "dev:electron": "cd dev && yarn 'dev:electron' && cd ..",
     "build": "node build/index.js",
-    "build:js": "node build/script.javascript.js"
+    "build:js": "node build/script.javascript.js",
+    "lint": "eslint --ext .js,.vue src && eslint --ext .js,.vue dev/src",
+    "lint:fix": "eslint --ext .js,.vue src --fix && eslint --ext .js,.vue dev/src --fix",
+    "stylelint": "stylelint {src,dev/src}/**/**/**/*.{vue,css,sass,scss}",
+    "stylelint:fix": "stylelint {src,dev/src}/**/**/**/*.{vue,css,sass,scss} --fix",
+    "lint:all": "npm run lint && npm run stylelint",
+    "lint:all:fix": "npm run lint:fix && npm run stylelint:fix",
+    "format": "prettier --write \"{src,dev/src}/**/*.{js,vue,scss}\""
   },
   "repository": {
     "type": "git",
@@ -23,40 +26,57 @@
   "bugs": "https://github.com/LinkageRetail/quasar-app-extension-boilerplate/issues",
   "homepage": "https://github.com/LinkageRetail/quasar-app-extension-boilerplate",
   "dependencies": {
-    "loadsh": "^0.0.4",
-    "vue": "2.6.14"
+    "lodash": "^4.17.21",
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0"
   },
   "devDependencies": {
-    "@quasar/app": "^2.0.0",
+    "@babel/eslint-parser": "^7.17.0",
+    "@quasar/app-webpack": "^3.0.0",
     "@quasar/extras": "^1.0.0",
     "@rollup/plugin-buble": "^0.21.3",
     "@rollup/plugin-commonjs": "^20.0.0",
     "@rollup/plugin-json": "^4.0.0",
     "@rollup/plugin-node-resolve": "^11.2.1",
     "@rollup/plugin-replace": "^2.4.2",
-    "@vue/compiler-sfc": "^3.2.31",
+    "@types/lodash": "^4.14.182",
+    "@types/node": "^12.20.21",
+    "@types/webpack-env": "^1.16.4",
+    "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "@typescript-eslint/parser": "5.10.0",
     "chalk": "^4.1.0",
     "core-js": "^3.6.5",
-    "eslint-webpack-plugin": "^2.4.0",
+    "eslint": "^8.10.0",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-vue": "^8.5.0",
     "fs-extra": "^8.1.0",
     "open": "^7.3.0",
     "postcss": "^8.0.0",
-    "quasar": "^1.0.0",
+    "prettier": "^2.5.1",
+    "quasar": "^2.0.0",
     "rimraf": "^3.0.0",
     "rollup": "^2.45.0",
     "rollup-plugin-postcss": "^4.0.0",
+    "rollup-plugin-typescript": "^1.0.1",
     "rollup-plugin-vue": "^6.0.0",
+    "stylelint": "^13.8.0",
+    "stylelint-config-prettier": "^8.0.2",
+    "stylelint-config-sass-guidelines": "^7.1.0",
+    "stylelint-config-standard": "^20.0.0",
+    "stylelint-order": "^4.1.0",
+    "stylelint-scss": "^3.18.0",
     "uglify-js": "^3.13.3",
     "zlib": "^1.0.5"
   },
   "browserslist": [
-    "last 4 Chrome versions",
-    "last 4 Firefox versions",
-    "last 2 Edge versions",
-    "last 4 Safari versions",
-    "last 4 Android versions",
-    "last 4 ChromeAndroid versions",
-    "last 4 FirefoxAndroid versions",
-    "last 4 iOS versions"
+    "last 10 Chrome versions",
+    "last 10 Firefox versions",
+    "last 4 Edge versions",
+    "last 7 Safari versions",
+    "last 8 Android versions",
+    "last 8 ChromeAndroid versions",
+    "last 8 FirefoxAndroid versions",
+    "last 10 iOS versions",
+    "last 5 Opera versions"
   ]
 }
