@@ -6,33 +6,35 @@
       subtitle="BACKEND MANAGEMENT SYSTEM"
       version="Version: 1.0.0"
       :model-type="type"
-      @login="login"
-      @sendEmail="sendEmail"
-      @resetPassword="resetPassword"
+      @login="onLogin"
+      @sendEmail="onSendEmail"
+      @resetPassword="onResetPassword"
     />
   </q-page>
 </template>
-<script>
-export default {
-  data() {
+
+<script lang="ts">
+import { defineComponent, ref } from 'vue';
+
+export default defineComponent({
+  setup() {
+    const type = ref('');
     return {
-      type: '',
+      type,
+      onLogin(form: { username: string; password: string }) {
+        console.log('onLogin');
+        console.log(form);
+      },
+      onSendEmail(email: string) {
+        console.log('onSendEmail');
+        console.log(email);
+        type.value = 'RESET';
+      },
+      onResetPassword(form: { newPassword: string; rePassword: string }) {
+        console.log('onResetPassword');
+        console.log(form);
+      },
     };
   },
-  methods: {
-    login(val) {
-      console.log('login');
-      console.log(val);
-    },
-    sendEmail(val) {
-      console.log('sendEmail');
-      console.log(val);
-      this.type = 'RESET';
-    },
-    resetPassword(val) {
-      console.log('resetPassword');
-      console.log(val);
-    },
-  },
-};
+});
 </script>
