@@ -1,51 +1,70 @@
-// Configuration for your app
-// https://quasar.dev/quasar-cli/quasar-conf-js
+/* eslint-env node */
 
-const ESLintPlugin = require('eslint-webpack-plugin');
+/*
+ * This file runs in a Node context (it's NOT transpiled by Babel), so use only
+ * the ES6 features that are supported by your Node version. https://node.green/
+ */
+
+// Configuration for your app
+// https://v2.quasar.dev/quasar-cli-webpack/quasar-config-js
+
+/* eslint-disable @typescript-eslint/no-var-requires */
+
+const { configure } = require('quasar/wrappers');
 const path = require('path');
 
-module.exports = function (ctx) {
+module.exports = configure(function (ctx) {
   return {
-    // https://v1.quasar.dev/quasar-cli/supporting-ts
-    supportTS: false,
+    // https://v2.quasar.dev/quasar-cli-webpack/supporting-ts
+    supportTS: {
+      tsCheckerConfig: {
+        eslint: {
+          enabled: true,
+          files: './src/**/*.{ts,tsx,js,jsx,vue}',
+        },
+      },
+    },
 
-    // https://v1.quasar.dev/quasar-cli/prefetch-feature
+    // https://v2.quasar.dev/quasar-cli-webpack/prefetch-feature
     // preFetch: true,
 
     // app boot file (/src/boot)
     // --> boot files are part of "main.js"
-    // https://v1.quasar.dev/quasar-cli/boot-files
-    boot: ['register.js'],
+    // https://v2.quasar.dev/quasar-cli-webpack/boot-files
+    boot: ['register'],
 
-    // https://v1.quasar.dev/quasar-cli/quasar-conf-js#Property%3A-css
+    // https://v2.quasar.dev/quasar-cli-webpack/quasar-config-js#Property%3A-css
     css: [],
 
     // https://github.com/quasarframework/quasar/tree/dev/extras
     extras: [
       // 'ionicons-v4',
-      // 'mdi-v4',
-      // 'fontawesome-v5',
+      // 'mdi-v5',
+      // 'fontawesome-v6',
       // 'eva-icons',
       // 'themify',
+      // 'line-awesome',
       // 'roboto-font-latin-ext', // this or either 'roboto-font', NEVER both!
-      // 'roboto-font' // optional, you are not bound to it
-      'material-icons', // optional, you are not bound to it
+
+      'material-icons',
       'material-icons-outlined',
       'material-icons-round',
       'material-icons-sharp',
     ],
 
-    // Full list of options: https://quasar.dev/quasar-cli/quasar-conf-js#Property%3A-build
+    // Full list of options: https://v2.quasar.dev/quasar-cli-webpack/quasar-config-js#Property%3A-build
     build: {
-      vueRouterMode: 'history',
+      vueRouterMode: 'history', // available values: 'hash', 'history'
+
       // transpile: false,
+      // publicPath: '/',
 
       // Add dependencies for transpiling with Babel (Array of string/regex)
       // (from node_modules, which are by default not transpiled).
       // Applies only if "transpile" is set to true.
       // transpileDependencies: [],
 
-      // rtl: false, // https://v1.quasar.dev/options/rtl-support
+      // rtl: true, // https://v2.quasar.dev/options/rtl-support
       // preloadChunks: true,
       // showProgress: false,
       // gzip: true,
@@ -54,27 +73,22 @@ module.exports = function (ctx) {
       // Options below are automatically set depending on the env, set them if you want to override
       // extractCSS: false,
 
-      // https://v1.quasar.dev/quasar-cli/handling-webpack
+      // https://v2.quasar.dev/quasar-cli-webpack/handling-webpack
       // "chain" is a webpack-chain object https://github.com/neutrinojs/webpack-chain
-      chainWebpack(chain) {
-        chain.plugin('eslint-webpack-plugin').use(ESLintPlugin, [{ extensions: ['js', 'vue'] }]);
-        chain.resolve.alias.merge({
-          ui: path.resolve(__dirname, `../src/index.esm.js`),
-        });
-      },
+      // chainWebpack(chain) { },
     },
 
-    // Full list of options: https://v1.quasar.dev/quasar-cli/quasar-conf-js#Property%3A-devServer
+    // Full list of options: https://v2.quasar.dev/quasar-cli-webpack/quasar-config-js#Property%3A-devServer
     devServer: {
-      https: false,
+      server: {
+        type: 'http',
+      },
       port: 8080,
       open: true, // opens browser window automatically
     },
 
-    // https://v1.quasar.dev/quasar-cli/quasar-conf-js#Property%3A-framework
+    // https://v2.quasar.dev/quasar-cli-webpack/quasar-config-js#Property%3A-framework
     framework: {
-      iconSet: 'material-icons', // Quasar icon set
-      lang: 'en-us', // Quasar language pack
       config: {
         notify: {
           type: 'positive',
@@ -92,20 +106,16 @@ module.exports = function (ctx) {
           /* brand defaults */
         },
         screen: {
-          // https://v1.quasar.dev/options/screen-plugin#introduction
+          // https://v2.quasar.dev/options/screen-plugin#introduction
           bodyClasses: true,
         },
       },
 
-      // Possible values for "importStrategy":
-      // * 'auto' - (DEFAULT) Auto-import needed Quasar components & directives
-      // * 'all'  - Manually specify what to import
-      importStrategy: 'auto',
+      iconSet: 'material-icons', // Quasar icon set
 
-      // https://v1.quasar.dev/layout/grid/introduction-to-flexbox
       cssAddon: true,
 
-      // For special cases outside of where "auto" importStrategy can have an impact
+      // For special cases outside of where the auto-import strategy can have an impact
       // (like functional components as one of the examples),
       // you can manually specify Quasar components/directives to be available everywhere:
       //
@@ -125,14 +135,28 @@ module.exports = function (ctx) {
     },
 
     // animations: 'all', // --- includes all animations
-    // https://v1.quasar.dev/options/animations
+    // https://v2.quasar.dev/options/animations
     animations: 'all',
 
+    // https://v2.quasar.dev/quasar-cli-webpack/developing-ssr/configuring-ssr
     ssr: {
+      pwa: false,
+
+      // manualStoreHydration: true,
+      // manualPostHydrationTrigger: true,
+
+      prodPort: 3000, // The default port that the production server should use
+      // (gets superseded if process.env.PORT is specified at runtime)
+
+      maxAge: 1000 * 60 * 60 * 24 * 30,
+      // Tell browser when a file from the server should expire from cache (in ms)
+
+      // chainWebpackWebserver (/* chain */) {},
+
       middlewares: [
         ctx.prod ? 'compression' : '',
         'render', // keep this as last one
       ],
     },
   };
-};
+});
