@@ -1,5 +1,7 @@
 <template>
-  <q-page padding>
-    <FileUpload label="Drop file here Or Select" api="" />
+  <q-page padding class="flex items-center justify-center">
+    <div style="width: 500px">
+      <FileUpload label="Drop file here Or Select" api="" />
+    </div>
   </q-page>
 </template>
