@@ -2,9 +2,10 @@
   <q-page padding>
     <FormDialog
       title="New Category"
-      :visible="formDialogVisible"
-      @on-cancel="cancelFormDialog"
-      @on-save="saveFormDialog"
+      :loading="loading"
+      @cancel="cancelFormDialog"
+      @save="saveFormDialog"
+      v-model="formDialogVisible"
     >
       <template #content> content </template>
     </FormDialog>
@@ -13,20 +14,25 @@
   </q-page>
 </template>
 
-<script>
-export default {
-  data() {
+<script lang="ts">
+import { defineComponent, ref } from 'vue';
+
+export default defineComponent({
+  setup() {
     return {
-      formDialogVisible: false,
+      loading: ref(false),
+      formDialogVisible: ref(false),
     };
   },
   methods: {
     cancelFormDialog() {
+      console.log('cancelFormDialog');
       this.formDialogVisible = false;
     },
     saveFormDialog() {
+      console.log('saveFormDialog');
       this.formDialogVisible = false;
     },
   },
-};
+});
 </script>
