@@ -1,5 +1,10 @@
 import { route } from 'quasar/wrappers';
-import VueRouter from 'vue-router';
+import {
+  createMemoryHistory,
+  createRouter,
+  createWebHashHistory,
+  createWebHistory,
+} from 'vue-router';
 
 import routes from './routes';
 
@@ -12,18 +17,25 @@ import routes from './routes';
  * with the Router instance.
  */
 
-export default route(({ Vue /* store, ssrContext */ }) => {
-  Vue.use(VueRouter);
+export default route(function (/* { store, ssrContext } */) {
+  const createHistory = process.env.SERVER
+    ? createMemoryHistory
+    : process.env.VUE_ROUTER_MODE === 'history'
+      ? createWebHistory
+      : createWebHashHistory;
 
-  const Router = new VueRouter({
+  const Router = createRouter({
     scrollBehavior: () => ({ left: 0, top: 0 }),
     routes,
-    // Leave these as is and change from quasar.conf.js instead!
+
+    // Leave this as is and make changes in quasar.conf.js instead!
     // quasar.conf.js -> build -> vueRouterMode
     // quasar.conf.js -> build -> publicPath
-    mode: process.env.VUE_ROUTER_MODE,
-    base: process.env.VUE_ROUTER_BASE,
+    history: createHistory(
+      process.env.MODE === 'ssr' ? void 0 : process.env.VUE_ROUTER_BASE
+    ),
   });
+
 
   // we get each page from server first!
   if (process.env.MODE === 'ssr' && process.env.CLIENT) {
