@@ -2,20 +2,15 @@
   <!-- Example: -->
   <!--  <FormDialog
       title="New Category"
-      :visible="formDialogVisible"
-      @on-cancel="cancelFormDialog"
-      @on-save="saveFormDialog"
+      :loading="loading"
+      @cancel="cancelFormDialog"
+      @save="saveFormDialog"
+      v-model="formDialogVisible"
     >
-      <template #content>
-      </template>
-    </FormDialog> 
+      <template #content> content </template>
+    </FormDialog>
   -->
-  <q-dialog
-    content-class="dialog"
-    persistent
-    :value="visible"
-    @hide="$emit('hide') && $emit('input', false)"
-  >
+  <q-dialog content-class="dialog" :persistent="persistent" @hide="onHide" v-model="model">
     <q-card class="dialog-card">
       <!-- Dialog Header -->
       <q-card-section class="dialog-header">
@@ -38,26 +33,47 @@
           no-caps
           class="q-mr-xs q-px-sm"
           color="blue"
+          :loading="loading"
           :label="saveLabel"
-          @click="$emit('on-save')"
+          @click="onSave"
         />
         <q-btn
           unelevated
           no-caps
           color="grey"
           class="q-px-sm"
+          :disable="loading"
           :label="cancelLabel"
-          @click="$emit('on-cancel')"
+          @click="onCancel"
         />
       </q-card-actions>
     </q-card>
   </q-dialog>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent, ref } from 'vue';
+import { QDialog, QCard, QCardSection, QCardActions, QBtn } from 'quasar';
+
+import { useModelWrapper } from '../hooks';
+
+export default defineComponent({
   name: 'FormDialog',
+  components: { QDialog, QCard, QCardSection, QCardActions, QBtn },
+  emits: ['modelValue', 'hide', 'input', 'save', 'cancel'],
   props: {
+    modelValue: {
+      type: Boolean,
+      default: false,
+    },
+    persistent: {
+      type: Boolean,
+      default: true,
+    },
+    loading: {
+      type: Boolean,
+      default: false,
+    },
     title: {
       type: String,
       default: 'Dialog',
@@ -70,13 +86,22 @@ export default {
       type: String,
       default: 'cancel',
     },
-    visible: {
-      type: Boolean,
-      default: false,
-      required: true,
-    },
   },
-};
+  setup(props, context) {
+    return {
+      model: ref(useModelWrapper(props, context.emit, 'modelValue')),
+      onHide() {
+        context.emit('hide');
+      },
+      onSave() {
+        context.emit('save');
+      },
+      onCancel() {
+        context.emit('cancel');
+      },
+    };
+  },
+});
 </script>
 
 <style lang="scss" scoped>
@@ -90,7 +115,7 @@ export default {
     padding-bottom: 2rem;
     height: 40px;
     color: #fff;
-    background-color: var(--q-color-primary);
+    background-color: var(--q-primary);
 
     > span {
       padding-bottom: 0.25rem;
