@@ -1,10 +1,10 @@
-<script>
+<script lang="ts">
+import { defineComponent, h } from 'vue';
 import { QSelect } from 'quasar';
-import _ from 'lodash';
 
-export default {
+export default defineComponent({
   name: 'FieldSelect',
-  extends: QSelect,
+  components: { QSelect },
   props: {
     dense: {
       type: Boolean,
@@ -48,7 +48,7 @@ export default {
     },
     rules: {
       type: Array,
-      default: () => [val => !_.isNull(val) && !_.isUndefined(val)],
+      default: () => [(val: any) => !!val],
     },
     options: {
       type: Array,
@@ -71,8 +71,20 @@ export default {
       default: 'value',
     },
   },
-};
+  emits: ['update:modelValue'],
+  setup(props, context) {
+    return () =>
+      /**
+       * @see https://vuejs.org/guide/extras/render-function.html#v-model
+       */
+      h(QSelect, {
+        ...props,
+        'onUpdate:modelValue': (value: any) => context.emit('update:modelValue', value),
+      });
+  },
+});
 </script>
+
 <style lang="scss" scoped>
 .q-field {
   padding-bottom: 0;
