@@ -1,10 +1,10 @@
-<script>
+<script lang="ts">
+import { defineComponent, h } from 'vue';
 import { QSelect } from 'quasar';
-import _ from 'lodash';
 
-export default {
+export default defineComponent({
   name: 'FilterSelect',
-  extends: QSelect,
+  components: { QSelect },
   props: {
     dense: {
       type: Boolean,
@@ -43,29 +43,16 @@ export default {
       default: 'value',
     },
   },
-  data() {
-    return {
-      optionsShadow: this.options,
-    };
-  },
-  methods: {
-    handleFilter(val, update) {
-      if (val === '') {
-        update(() => {
-          this.options = this.optionsShadow;
-          // with Quasar v1.7.4+
-          // here you have access to "ref" which
-          // is the Vue reference of the QSelect
-        });
-        return;
-      }
-      update(() => {
-        const needle = val.toLowerCase();
-        this.options = this.optionsShadow.filter(
-          v => _.get(v, this.optionLabel).toLowerCase().indexOf(needle) > -1
-        );
+  emits: ['update:modelValue'],
+  setup(props, context) {
+    return () =>
+      /**
+       * @see https://vuejs.org/guide/extras/render-function.html#v-model
+       */
+      h(QSelect, {
+        ...props,
+        'onUpdate:modelValue': (value: any) => context.emit('update:modelValue', value),
       });
-    },
   },
-};
+});
 </script>
