@@ -113,9 +113,15 @@
   </BaseLayout>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent } from 'vue';
+import { QCard, QCardSection } from 'quasar';
+
+import BaseLayout from './BaseLayout.vue';
+
+export default defineComponent({
   name: 'DetailLayout',
+  components: { BaseLayout, QCard, QCardSection },
   props: {
     title: {
       type: String,
@@ -138,10 +144,7 @@ export default {
       default: () => ({ md: [9, 3], lg: [8, 4] }),
     },
   },
-  components: {
-    BaseLayout: () => import('./BaseLayout'),
-  },
-};
+});
 </script>
 
 <style lang="scss" scoped>
