@@ -121,9 +121,15 @@
   </BaseLayout>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent } from 'vue';
+import { QCard, QCardSection } from 'quasar';
+
+import BaseLayout from './BaseLayout.vue';
+
+export default defineComponent({
   name: 'DetailListLayout',
+  components: { BaseLayout, QCard, QCardSection },
   props: {
     title: {
       type: String,
@@ -134,10 +140,7 @@ export default {
       default: true,
     },
   },
-  components: {
-    BaseLayout: () => import('./BaseLayout'),
-  },
-};
+});
 </script>
 
 <style lang="scss" scoped>
